APEG.State

Description

This module contains the definition of an APEG state used throughout the interpreter and type-checker. The abstract data type allows for consulting and updating value and type environments, obtain the string currently accepted by the recognizer, setting and consulting and transforming the results and finally manipulation of the input.

Documentation

type APegTuple a = (VEnv, TyEnv, TraceStr, Input, Result a)

type Result a = Either [String] a

The result buit by a parser. Can either be an list of error messages or an actual value of some type a.

type APegSt a = State (APegTuple a)

type VEnv = Map String Value

Value environment, a mapping form names to Values

type TyEnv = Map String (Type, TyRuleEnv)

Type environment, a mapping from names to a non terminal Type an it's inner enviroment

type TyRuleEnv = Map String Type

Type environment for a rule. Maps varialble names to

type TraceStr = Maybe String

A TraceStr is a String that may be initialized, an actual string, or may not be initialized, a null String. Any operation on null String will result a null String. This definition is for the sake of memory usage on the APEG bind operation, that migth require toe keep the whole input as the recognized input.

type Input = String

The recognizer input is represented by an String.

rVal :: a -> Result a

Builds a success Result from a particular value.

rErr :: String -> Result a

Builds a failure Result from an error message.

(?:) :: Char -> TraceStr -> TraceStr

nullTraceStr :: TraceStr

emptyTraceStr :: TraceStr

errorTuple :: String -> APegTuple a -> APegTuple a

isOkTuple :: APegTuple a -> Bool

valEnv :: APegTuple a -> VEnv

tyEnv :: APegTuple a -> TyEnv

upValEnv :: (VEnv -> VEnv) -> APegTuple a -> APegTuple a

upTyEnv :: (TyEnv -> TyEnv) -> APegTuple a -> APegTuple a

withResult :: ([String] -> Result b) -> (a -> Result b) -> APegTuple a -> APegTuple b

valIsMExpr :: Value -> Bool

valIsMPeg :: Value -> Bool

strVal :: Value -> String

expFromVal :: Value -> Expr

varNameFromVal :: Value -> Var

apegFromVal :: Value -> APeg

zeroSt :: ApegGrm -> String -> APegTuple ()

ruleEntry :: ApegRule -> (NonTerminal, (Type, TyRuleEnv))

isOk :: APegSt () Bool